:recycle: [Refactor] types 파일 분리 #main
