+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router } from "react-router-dom"
 import Header from "./entities/ui/layout/Header.tsx"
 import Footer from "./entities/ui/layout/Footer.tsx"
-import { PostsManagerPage } from "./pages/ui/manager/PostManagerPage.tsx"
+import { PostManagerPage } from "./pages/ui/manager/PostManagerPage.tsx"
 
 const App = () => {
   return (
@@ -9,7 +9,7 @@ const App = () => {
       <div className="flex flex-col min-h-screen">
         <Header />
         <main className="flex-grow container mx-auto px-4 py-8">
-          <PostsManagerPage />
+          <PostManagerPage />
         </main>
         <Footer />
       </div>
