+++ src/shared/types/tag.ts
@@ -1,0 +1,6 @@
+interface Tag {
+  url: string
+  slug: string
+}
+
+export type { Tag }
