+++ src/shared/types/url.ts
@@ -1,0 +1,10 @@
+interface URLParams {
+  skip?: number
+  limit?: number
+  search?: string
+  sortBy?: string
+  sortOrder?: string
+  tag?: string
+}
+
+export type { URLParams }
